[makeotfParts] decode glyphName to (unicode) str
`glyphName` is first encoded to bytes, and then each character tested to see
if they are in the set of `_validCharacters`. However, the latter is made of
`str` instances. This means `c in _validCharacters` is always returning False,
and the resulting glyphName is left as an empty string.
This in turn makes `isLegalGlyphName` function fail with IndexError:

```
Traceback (most recent call last):
  File "ufo2fdk/Lib/ufo2fdk/makeotfParts.py", line 545, in isLegalGlyphName
    if glyphName[0] in _digits:
```